Rename script.ts to config.ts
